Moves and renames to clean up the file structure.
