+++ tools/lcsu/config/callslipQuery.inc.php
